Add a happy-path test
