+++ fixtures/json-artifact.json
@@ -1,0 +1,1 @@
+{"some": "json"}
