+++ fixtures/xml-artifact.xml
@@ -1,0 +1,1 @@
+<some-xml></some-xml>
